Created header files for functions and class
